+++ src/overlays/koopa/3C27C0.c
@@ -513,7 +513,7 @@ INCLUDE_ASM("asm/nonmatchings/overlays/koopa/3C27C0", func_801092A0_3C60C0_koopa
 
 s32 func_801092D4_3C60F4_koopa(void) {
     Hu3DCamInit(1);
-    func_800E6630_FA250_shared_board(hvq_data_ROM_START);
+    func_800E6630_FA250_shared_board((u32)hvq_data_ROM_START);
     MBBackLoad(2);
     return func_800E9344_FCF64_shared_board(100.0f, 10000.0f);
 }
