+++ src/overlays/opboardtutorial/3CDEC0.c
@@ -46,7 +46,7 @@ INCLUDE_ASM("asm/nonmatchings/overlays/opboardtutorial/3CDEC0", func_8010751C_3C
 void func_80107550_3CFA70_opboardtutorial(void) {
     Hu3DCamInit(1);
     HmfLightColorSet(1, 0xFF, 0xFF, 0xFF);
-    func_800E6630_FA250_shared_board(hvq_data_ROM_START);
+    func_800E6630_FA250_shared_board((u32)hvq_data_ROM_START);
     MBBackLoad(0x15);
 }
